Please have a read.
The biggest change I made here is Text Class being able to create, search (can do multiple times on the same dataset), modify and delete. I hope this becomes a good foundation for other classes.
Like John said, I merged a constructor with a create().
Changed pimItems to double Map so that we can simply use ID as a key to delete and modify with O(1).