Export & Load bug solved
The problem of date became null was solved, I suppose it was indeed caused by the issue of static. Meanwhile, I noticed that the static field "next id" of PIR classes and the nested hashmap of pirInterface caused some problems: when the program started, the next id fields of PIR classes are initialized to 1. So when data is loaded to the program, those next id fields are not being updated, resulting that create new PIRs will replace some of the existing PIRs loaded from the .pim file. I fixed it based without changing too much on the structure, but if you guys have other better approach, feel free to adopt it